Now iupred works

+++ arne/proks_vs_euks/run_iupred.py
@@ -4,7 +4,7 @@ import commands
 
 
 
-input_dir = "../results/uniprot_sequences_groups/"
+input_dir = "/pfs/nobackup/home/w/wbasile/proks_euks/results/uniprot_sequences_groups/"
 
 #~ file_list = []
 #~ for f in os.listdir(input_dir):
@@ -23,16 +23,26 @@ for f in os.listdir(input_dir):
         
 
 for f in file_list:
-    
+    print f
     for iupred_param in ["long", "short"]:
         iupred_data_file = f +  ".data_iupred_" + iupred_param 
+        iupred_data_file_04 = f +  ".data_iupred_0.4_" + iupred_param 
         iupred_data_file_raw = f +  ".data_iupred_"+iupred_param+"_raw" 
         
+        # This makes it possible to run in parallell
+        if os.path.exists(iupred_data_file) and  os.path.exists(iupred_data_file_04):
+            print "Skipping ", f
+            continue
         # clean empty files
-        #~ if os.path.exists(iupred_data_file):
-            #~ st = os.stat(iupred_data_file)
-            #~ if st.st_size == 0:
-                #~ os.system("rm " + iupred_data_file)
+        #if os.path.exists(iupred_data_file):
+        #    st = os.stat(iupred_data_file)
+        #    if st.st_size == 0:
+        #        os.system("rm " + iupred_data_file)
+        #
+        #if os.path.exists(iupred_data_file_04):
+        #    st = os.stat(iupred_data_file_04)
+        #    if st.st_size == 0:
+        #        os.system("rm " + iupred_data_file_04)
 
 
         if not os.path.exists(iupred_data_file):
@@ -43,6 +53,34 @@ for f in file_list:
         
             if os.path.exists(iupred_data_file_raw):
 
+                # remove empty files
+                st = os.stat(iupred_data_file_raw)
+                if st.st_size == 0:
+                    os.system("rm " + iupred_data_file_raw)
+                #else:
+                #    group_num = int(f.split(".fasta")[0].split("_")[-1])
+                    
+                    # remove the incomplete raw files
+                    if group_num != 1208:
+                        cmd = "grep -c Prediction " + f 
+                        ret = int(commands.getoutput(cmd))
+                        
+                        if ret != 50000:
+                            print ("rm " + iupred_data_file_raw)
+                            sys.exit()
+                            os.system("rm " + iupred_data_file_raw)
+
+            if not os.path.exists(iupred_data_file_raw):
+            
+                cmd = "./iupred_multi " + f + " " + iupred_param + " > " + iupred_data_file_raw
+                print cmd
+                os.system(cmd)
+        if not os.path.exists(iupred_data_file_04):
+            print "creating " + iupred_data_file_04
+            os.system("touch " + iupred_data_file_04)
+            if os.path.exists(iupred_data_file_raw):
+
+                # remove empty files
                 st = os.stat(iupred_data_file_raw)
                 if st.st_size == 0:
                     os.system("rm " + iupred_data_file_raw)
@@ -51,26 +89,31 @@ for f in file_list:
                     
                     # remove the incomplete raw files
                     if group_num != 1208:
-                        cmd = "grep Prediction " + f + " | wc -l"
+                        cmd = "grep -c Prediction " + f 
                         ret = int(commands.getoutput(cmd))
                         
                         if ret != 50000:
-                            os.system("rm " + iupred_data_file_raw)
-                            
-                        
-            
+                               print ("rm " + iupred_data_file_raw)
+                               sys.exit()
+                               os.system("rm " + iupred_data_file_raw)
+
             if not os.path.exists(iupred_data_file_raw):
             
                 cmd = "./iupred_multi " + f + " " + iupred_param + " > " + iupred_data_file_raw
                 print cmd
                 os.system(cmd)
+        
+                            
+                        
+            
                 
             #PARSE
-            print "parsing " + iupred_data_file_raw
+            print "parsing " + iupred_data_file_raw + " into " + iupred_data_file
             ps = filter(None, open(iupred_data_file_raw).read().split("# Prediction output "))
 
             iupred_dic = {}
             for p in ps:
+                #print p
                 query_id = ""
                 values = []
 
@@ -98,5 +141,38 @@ for f in file_list:
                 for k in iupred_dic.keys():
                     outf.write(">" + k + "\n" + iupred_dic[k] + "\n")
 
+            #PARSE
+            print "parsing " + iupred_data_file_raw  + " into " + iupred_data_file
+            ps = filter(None, open(iupred_data_file_raw).read().split("# Prediction output "))
 
+            iupred_dic = {}
+            for p in ps:
+                query_id = ""
+                values = []
+
+                lines = filter(None,p.split("\n"))
+                for line in lines:
+                    if line[0] == "#":
+                        query_id = line.split()[1]
+                    else:
+                        values += [float(line.split()[-1])]
+
+                # calculate the disorder as the fraction of disordered (>0.5) residues
+                if query_id != "":
+                    
+                    diso_aa = ""
+                    for a in values:
+                        if a > 0.4:
+                            diso_aa += "x"
+                        else:
+                            diso_aa += "n"
+                    
+                    iupred_dic[query_id] = diso_aa
+                    
+                    
+            with open(iupred_data_file_04, "w") as outf:
+                for k in iupred_dic.keys():
+                    outf.write(">" + k + "\n" + iupred_dic[k] + "\n")
+        #sys.exit()
+
